+++ src/main/java/com/example/servicemanagement/service/TenantService.java
@@ -11,8 +11,7 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
-import static com.example.servicemanagement.constant.Constant.ERR_INSERT_DUPLICATE_TENANT;
-import static com.example.servicemanagement.constant.Constant.ERR_UPDATE_INVALID_TENANT;
+import static com.example.servicemanagement.constant.Constant.*;
 
 @Service
 public class TenantService {
@@ -34,7 +33,7 @@ public class TenantService {
     }
 
     public List<Tenant> getAllByApartmentId(Integer apartmentId) {
-        return this.tenantRepository.findTenantsByApartmentId(apartmentId);
+        return this.tenantRepository.findTenantsByApartmentIdOrderByRoomNoAsc(apartmentId);
     }
 
     public void update(Integer id, Tenant body) {
@@ -46,10 +45,8 @@ public class TenantService {
 
         Tenant tenant = this.tenantRepository.findTenantById(id);
         tenant.setRoomNo(body.getRoomNo());
-        User user = tenant.getUser();
-        user.setName(body.getUser().getName());
-        user.setPhoneNo(body.getUser().getPhoneNo());
-        user.setPassword(body.getUser().getPassword());
+
+        this.userService.update(body.getUser().getId(), body.getUser());
 
         this.tenantRepository.saveAndFlush(tenant);
     }
@@ -70,6 +67,13 @@ public class TenantService {
 
     public void delete(Integer id) {
         Tenant tenant = this.tenantRepository.findTenantById(id);
+
+        boolean canDelete = this.tenantRepository.checkCanDelete(tenant.getUser().getId());
+
+        if (!canDelete) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, ERR_DELETE_INVALID_TENANT);
+        }
+
         this.tenantRepository.deleteById(id);
         this.userService.delete(tenant.getUser().getId());
     }
